fix(osadapter): move argv/label helpers to untagged file (unbreaks Linux)
origin/master (#50) shipped with sameRoster/rosterFromArgv/workdir/interval
helpers in the darwin-only ctl_darwin.go while the untagged tests use them,
so the non-darwin build is red (undefined: sameRoster). Re-applies the move
to an untagged argv.go. darwin+linux+windows green.
